+++ src/lib/assets/Container.tsx
@@ -1,6 +1,6 @@
 import PubSub from '@harryy/pubsub'
 import { Trans } from '@lingui/macro'
-import { Box, Theme, Tooltip } from '@material-ui/core'
+import { Theme } from '@material-ui/core'
 import { makeStyles } from '@material-ui/styles'
 import React from 'react'
 import { DndContainerItem, RenderProps } from '../types'
@@ -82,6 +82,19 @@ const defaultContainer: DndContainerItem = {
                 ? { backgroundImage: `url(${state.style.backgroundImage})` }
                 : {})
         }
+
+        if (state?.outlook) {
+            return (
+                <table style={{ width: '100%' }}>
+                    <tbody>
+                        <tr>
+                            <td style={modifiedStyle}>{children}</td>
+                        </tr>
+                    </tbody>
+                </table>
+            )
+        }
+
         return (
             <div onClick={handleClick} style={modifiedStyle}>
                 {children}
@@ -99,6 +112,21 @@ const defaultContainer: DndContainerItem = {
                 ? { backgroundImage: `url(${state.style.backgroundImage})` }
                 : {})
         }
+
+        if (state?.outlook) {
+            return ` 
+                <table style="width: 100%">
+                    <tbody>
+                        <tr>
+                            <td style="${styleToCss(modifiedStyle)}">
+                                ${children}
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+            `
+        }
+
         return `
             <div style="${styleToCss(modifiedStyle)}">
                 ${children}
@@ -115,7 +143,8 @@ const defaultContainer: DndContainerItem = {
             width: '100%',
             boxSizing: 'border-box',
             textAlign: 'center'
-        }
+        },
+        outlook: false
     },
     settings: [
         {
@@ -129,7 +158,8 @@ const defaultContainer: DndContainerItem = {
                     type: 'backgroundColor',
                     grid: 12,
                     hideIfSet: 'style.backgroundImage'
-                }
+                },
+                { id: 'outlook', type: 'labeledSwitch', grid: 12, label: <Trans>Outlook</Trans> }
             ],
             label: <Label />,
             type: 'container'