+++ src/lib/assets/components/FileUpload.tsx
@@ -1,21 +1,33 @@
+import { FormikValues, useFormikContext } from 'formik'
 import React from 'react'
 import PubSub from '@harryy/pubsub'
 import { Trans } from '@lingui/macro'
-import { Grid, FormHelperText, InputLabel } from '@material-ui/core'
+import { Grid, FormHelperText, FormLabel, Link } from '@material-ui/core'
 import { DndComponentItem, RenderProps } from '../../types'
-import Uploader from '../../components/preferences/components/Uploader'
+import { getComponentState, getFormikProps } from '../../utils'
+import Uploader from '../libs/Uploader'
 
 export default {
-    render: (renderProps: RenderProps, id: string) => {
-        if (!renderProps.item || !id) {
-            return null
-        }
+    render: (renderProps: RenderProps, id: string, formKey: string) => {
         const handleClick = (ev: React.MouseEvent) => {
-            ev.preventDefault()
-            PubSub.publish('component/click', { type: 'form-elements', data: id })
+            if (renderProps.buildermode) {
+                ev.preventDefault()
+                PubSub.publish('component/click', { type: 'form-elements', data: id })
+            }
         }
-        const state = renderProps.state.entities[renderProps.item.id]?.values?.[id]
+        const state = getComponentState(renderProps, id)
         const labelText = `${state?.question}${state?.required ? '*' : ''}`
+        let formikProps: any = {}
+        const formik = useFormikContext<FormikValues>()
+        if (formik && formKey) {
+            formikProps = getFormikProps(formKey, formik, (value) => {
+                return {
+                    text: value,
+                    valueType: 'String'
+                }
+            })
+            formikProps.value = formikProps.value?.text
+        }
         return (
             <Grid
                 container
@@ -24,11 +36,18 @@ export default {
                 onClick={handleClick}
             >
                 <Grid item xs={12}>
-                    <InputLabel htmlFor="my-input">{labelText}</InputLabel>
+                    <FormLabel component="legend" style={{ marginBottom: 4, display: 'block' }}>
+                        {labelText}
+                    </FormLabel>
                 </Grid>
-                <Uploader value="" onChange={() => {}} />
+
+                <Uploader
+                    id={`${renderProps?.item?.id}-${id}`}
+                    accept={state.accept}
+                    {...formikProps}
+                />
                 <Grid item xs={12}>
-                    <FormHelperText id="my-helper-text">{state?.hint}</FormHelperText>
+                    <FormHelperText>{state?.hint}</FormHelperText>
                 </Grid>
             </Grid>
         )
@@ -40,21 +59,32 @@ export default {
         question: 'Upload',
         hint: '',
         className: '',
-        multiple: false,
+        accept: 'image/*',
         required: false,
-        enabled: true
+        enabled: true,
+        itemType: 'File'
     },
     settings: [
         { id: 'question', type: 'labeledTextInput', grid: 12, label: <Trans>Question</Trans> },
         { id: 'hint', type: 'labeledTextInput', grid: 12, label: <Trans>Hint</Trans> },
         { id: 'className', type: 'labeledTextInput', grid: 12, label: <Trans>Class name</Trans> },
-        { id: 'required', type: 'labeledSwitch', grid: 12, label: <Trans>Required</Trans> },
         {
-            id: 'multiple',
-            type: 'labeledSwitch',
+            id: 'accept',
+            type: 'labeledTextInput',
             grid: 12,
-            label: <Trans>Allow multiple files</Trans>
+            label: <Trans>Accept</Trans>,
+            helperText: (
+                <>
+                    <Link
+                        href="https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/accept"
+                        target="_blank"
+                    >
+                        Read more about what patterns work
+                    </Link>
+                </>
+            )
         },
+        { id: 'required', type: 'labeledSwitch', grid: 12, label: <Trans>Required</Trans> },
         { id: 'enabled', type: 'labeledSwitch', grid: 12, label: <Trans>Enabled</Trans> }
     ]
 } as DndComponentItem