+++ src/lib/components/DndItemPreview.tsx
@@ -10,9 +10,9 @@ import {
 import { makeStyles } from '@material-ui/styles'
 import clsx from 'clsx'
 import { bindHover, bindPopper, usePopupState } from 'material-ui-popup-state/hooks'
-import React from 'react'
+import React, { useEffect } from 'react'
 import PubSub from '@harryy/pubsub'
-import { DndBlockItem, RenderProps, SettingItemType } from '../types'
+import { DndBlockItem, RenderProps } from '../types'
 import { conditionBuilder, removeItem } from '../utils'
 
 interface Props extends RenderProps, React.HTMLAttributes<HTMLDivElement> {}
@@ -89,7 +89,8 @@ const DndItemPreview: React.FC<Props> = React.forwardRef<HTMLDivElement, Props>(
     ) => {
         const popupState = usePopupState({
             variant: 'popper',
-            popupId: `${item?.id}-actions`
+            popupId: `${item?.id}-actions`,
+            disableAutoFocus: true
         })
         const renderProps = {
             setState,